Move tx observation to hydra-tx
This allows the observation types to be easier re-used.
