+++ hydra-chain-observer/hydra-chain-observer.cabal
@@ -105,6 +105,7 @@ test-suite tests
     , hydra-node
     , hydra-prelude
     , hydra-test-utils
+    , hydra-tx
     , hydra-tx:testlib
     , QuickCheck
