+++ hydra-tx/test/Hydra/Tx/Contract/Close/CloseUsed.hs
@@ -26,7 +26,8 @@ import Hydra.Tx (
   registryUTxO,
   signatures,
  )
-import Hydra.Tx.Close (OpenThreadOutput (..), closeTx)
+import Hydra.Tx.Close (closeTx)
+import Hydra.Tx.CollectCom (OpenThreadOutput (..))
 import Hydra.Tx.Contract.Close.Healthy (
   healthyCloseLowerBoundSlot,
   healthyCloseUTxOHash,
