+++ hydra-tx/test/Hydra/Tx/Contract/Contest/Healthy.hs
@@ -13,7 +13,8 @@ import Hydra.Ledger.Cardano.Time (slotNoToUTCTime)
 import Hydra.Plutus.Extras (posixFromUTCTime)
 import Hydra.Plutus.Orphans ()
 import Hydra.Tx (registryUTxO)
-import Hydra.Tx.Contest (ClosedThreadOutput (..), contestTx)
+import Hydra.Tx.Close (ClosedThreadOutput (..))
+import Hydra.Tx.Contest (contestTx)
 import Hydra.Tx.ContestationPeriod (ContestationPeriod, fromChain)
 import Hydra.Tx.Crypto (HydraKey, MultiSignature, aggregate, sign)
 import Hydra.Tx.HeadId (mkHeadId)
